Migrate all the code for #[derive(Butcher)] in a separate module
This will allow to separate the code a bit better for structs/enums
